fix: switch to buildjet cache for e2e snapshot tests
Maintaining the Haskell's node database as cache has really demonstrated
to be unpractical. In an ideal world, we could just run Amaru against a
remote cardano-node; but we aren't there quite yet in terms of
networking.

So we still have to run a local cardano-node to connect to, which
requires a database; thus causing two issues:

- We need to store and restore that database on-demand; which is what
  we've been using caches for until now. Yet, we are limited to 10GB of
  caches on Github's free-tier; and it is cumbersome to ensure that our
  precious node database doesn't get evicted in favor of other caches.

- We're also hitting disk-space quota limit (14GB on free-tier runners)
  when running the full preprod e2e.

So, this commit:

- externalizes the cardano-node's db cache to buildjet which should avoid clashes
  between the node's database and the build caches.

- we conditionally use a buildjet runner for long-running e2e tests when
  merging to main; which comes with more disk space.

Signed-off-by: KtorZ <matthias.benkort@gmail.com>